+++ packages/primitives/src/components/TextInput/index.ts
@@ -1,0 +1,2 @@
+export { default as TextInput } from './textInput';
+export * from './textInput.types';
