+++ packages/primitives/src/components/TextInput/textInput.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import TextInput from '.';
+import { TextInput } from '.';
 
 export default {
   title: 'Primitives/TextInput',
